Now put food,reward,bomb objects into separate threads
however, found that the screen timeLeft for reward and bomb objects
did not update at all. The problem is that part of the reward/bomb object
is in the main thread (in the game) and the other part is in the separate
threads (i.e. th2/th3). so the part in the main thread remains the same
even though the data in the separate threads are updating.